+++ Changelog.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+#### 1.2.O (2017-03-15)
+
+* Separate the twig extension configuration from the PaginationManager to an OkaPaginationExtension class in Namespace Oka\PaginationBundle\Twig\OkaPaginationExtension
+* Allows you to enable or disable the twig extension
+* Allows to activate or deactivate the injection of globals variables in the twig template
+* Added params context in pagintaion template
+
 #### 1.1.1 (2017-02-16)
 
 * Added itemOffset in pagination resultset